generated_tasks/deathclock__actuarial-life-expectancy/environment/repo/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
groq>=0.13.0
